Address PR comments re: adding a test case & moving a function
